batch script for windows users
This script simplifies the process of activating your venv python environment then running the harvester.
It assumes you put the harvester in %USERPROFILE%\Documents\COUNTER_Harvester\C5.1_Harvester
where %USERPROFILE% is your Windows account's home directory (the folder that you see Desktop, Downloads, etc. as folder inside it).
If you chose a different location, just edit the PROJECT_DIR line at the top of the script to match where you put it. Don't add the "/src" part of the path because that's handled further down in the script.